requirements.txt
pinecone
openai
python-dotenv
pyTelegramBotAPI
haystack-ai
pinecone-haystack
sentence-transformers
requests
Pillow
docling
docling-haystack
